week03-homework/pyproject.toml
[project]
name = "week03-homework"
version = "0.1.0"
description = "Week 3 Homework - RAG and OCR"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv",
    "llama-index",
    "llama-index-core",
    "llama-index-llms-openai-like",
    "llama-index-llms-dashscope",
    "llama-index-embeddings-dashscope",
    "spacy", 
]
